Refactor to one FactCommand
