+++ commands/PhpFact/Implementations/CommandFactory.php
@@ -7,7 +7,6 @@ use SoerBot\Commands\PhpFact\Exceptions\CommandNotFoundException;
 use SoerBot\Commands\PhpFact\Implementations\Commands\FactCommand;
 use SoerBot\Commands\PhpFact\Implementations\Commands\ListCommand;
 use SoerBot\Commands\PhpFact\Implementations\Commands\StatCommand;
-use SoerBot\Commands\PhpFact\Implementations\Commands\FactExtendedCommand;
 
 class CommandFactory
 {
@@ -16,14 +15,8 @@ class CommandFactory
         if (preg_match('/^(?<command>[a-z]+)(?:\s+(?<position>\d+))?$/iSu', $input, $match)) {
             array_shift($match);
 
-            if (!empty($match['position']) && $match['command'] === 'fact') {
-                $position = $match['position'];
-
-                return new FactExtendedCommand($facts, ['position' => $position]);
-            }
-
             if ('fact' === $match['command']) {
-                return new FactCommand($facts);
+                return new FactCommand($facts, $match);
             }
 
             if ('stat' === $match['command']) {
